requirements.txt
streamlit
pillow
lxml
python_dotenv
unstructured[all-docs]
tiktoken
langchain
langchain-community
langchain-groq
langchain-core
langchain-google-genai
google-generativeai
google-genai
chromadb
matplotlib
ipython
notebook
jupyter